refactor(agent): deepen durable conversation lifecycle
